Compile with jsx using logic from before jsx
